pubsub-to-bigquery/producer/requirements.txt
google-cloud-pubsub==2.13.4
Faker==13.15.1
